Implement the ChangesTracking only for ToDos
