+++ opengever/base/configure.zcml
@@ -399,7 +399,4 @@
 
   <adapter factory=".response.ResponseContainer" />
 
-  <adapter factory=".response.NullChangesTracker" />
-  <adapter factory=".response.AutoResponseChangesTracker" />
-
 </configure>
